requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
asgiref==3.2.10
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
discord.py @ git+https://github.com/Rapptz/discord.py@b77af199392209017e59eff366b46354fa210f64
dnspython==2.0.0
gast==0.3.3
grpcio==1.29.0
h5py==2.10.0
idna==2.9
joblib==0.15.1
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.1
multidict==4.7.6
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.4
Pillow==7.1.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.0.0.dev6
pymongo==3.11.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.14
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sqlparse==0.3.1
threadpoolctl==2.1.0
tqdm==4.46.0
typing-extensions==3.7.4.2
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.5.0
